+++ lib/flutter_common.dart
@@ -26,6 +26,7 @@ export 'widget/common_load_more_view.dart';
 export 'widget/common_loading_view.dart';
 export 'widget/common_network_image_view.dart';
 export 'widget/common_slide_view.dart';
+export 'widget/common_toast.dart';
 export 'widget/common_web_view.dart';
 export 'widget/common_navigator_bar.dart';
 export 'common_constant.dart';