+++ lib/widget/common_toast.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+import 'package:fluttertoast/fluttertoast.dart';
+
+class CommonToast{
+
+  static void show(String msg, {
+    ToastGravity gravity = ToastGravity.CENTER,
+    Color backgroundColor = Colors.black87,
+    Color textColor = Colors.white,
+    double fontSize = 16.0,
+    Toast toastLength = Toast.LENGTH_SHORT,
+    int time = 2,
+  }){
+    Fluttertoast.showToast(
+        msg: msg,
+        toastLength: toastLength,
+        gravity: gravity,
+        timeInSecForIos: time,
+        backgroundColor: backgroundColor,
+        textColor: textColor,
+        fontSize: fontSize
+    );
+  }
+
+
+}